Typescript type fixes
