+++ frontend/app/types.d.ts
@@ -1,8 +1,35 @@
-import { type Namespace } from 'i18next';
+import { type Namespace, type ParseKeys } from 'i18next';
 
-import type gcweb from '~/public/locales/en/gcweb.json';
+import type common from '../public/locales/en/common.json';
+import type gcweb from '../public/locales/en/gcweb.json';
+import { type ClientEnv } from '~/utils/env.server';
+
+export type I18nResources = {
+  common: typeof common;
+  gcweb: typeof gcweb;
+};
+
+export type RouteHandleBreadcrumb = {
+  i18nKey: ParseKeys<keyof I18nResources>;
+  to?: string;
+};
+
+declare global {
+  interface Window {
+    env: ClientEnv;
+  }
+}
+
+/**
+ * @see https://www.i18next.com/overview/typescript
+ */
+declare module 'i18next' {
+  interface CustomTypeOptions {
+    resources: I18nResources;
+  }
+}
 
 export interface RouteHandle extends Record<string, unknown> {
   i18nNamespaces?: Namespace;
-  breadcrumbs?: Array<{ i18nKey: typeof gcweb; to?: string }>;
+  breadcrumbs?: Array<RouteHandleBreadcrumb>;
 }
